examples/bitcoin-price-tracker-0.7.1-aws-athena/requirements.txt
boto3>=1.34.0
requests>=2.32.5
pyyaml>=6.0.1
pyarrow>=14.0.0
